+++ internal/config/example.go
@@ -9,13 +9,31 @@ import (
 )
 
 func GenerateExampleYamlConfig(logger *zap.Logger, w io.StringWriter) error {
+	var variables []Variable
+
+	variableA := Variable{
+		Key:      "SHELL",
+		Value:    "/bin/bash",
+		Operator: ImmediateAssignment,
+	}
+	variables = append(variables, variableA)
+
+	variableB := Variable{
+		Key:      ".SHELLFLAGS",
+		Value:    "-eu -o pipefail -c",
+		Operator: ImmediateAssignment,
+	}
+	variables = append(variables, variableB)
+
 	tA := Target{
+		Name:            "targetA",
 		HelpDescription: "targetA just prints an output",
 		Commands:        []string{"echo \"Hello World\"", "echo \"This is how you specify commands!\""},
 		Display:         false,
 	}
 
 	tB := Target{
+		Name:            "targetB",
 		HelpDescription: "targetB just prints an output",
 		PreTargets:      []string{"targetA"},
 		Commands:        []string{"echo \"This is targetB!\"", "echo \"How are you doing?\""},
@@ -23,19 +41,29 @@ func GenerateExampleYamlConfig(logger *zap.Logger, w io.StringWriter) error {
 	}
 
 	config := Config{
-		Targets: map[string]Target{
-			"targetA": tA,
-			"targetB": tB,
-		},
+		Variables: variables,
+		Targets:   []Target{tA, tB},
 		Templates: OptionalTemplates{
 			HelpTemplate{
 				Enabled:   true,
 				Delimiter: "##",
 			},
-			VersionTemplate{
-				Enabled:    true,
-				Library:    "example",
-				MinVersion: "0.1.0",
+		},
+		Functions: OptionalFunctions{
+			RequireToolFunction{
+				Enabled: true,
+			},
+			RequireEnvFunction{
+				Enabled: true,
+			},
+			RequireFileFunction{
+				Enabled: true,
+			},
+			RequireDirFunction{
+				Enabled: true,
+			},
+			RequireConfirmationFunction{
+				Enabled: true,
 			},
 		},
 	}